+++ crates/eww/src/app.rs
@@ -1,5 +1,4 @@
 use crate::{
-    config,
     daemon_response::DaemonResponseSender,
     display_backend::DisplayBackend,
     error_handling_ctx,
@@ -17,12 +16,14 @@ use codespan_reporting::files::Files;
 use eww_shared_util::{Span, VarName};
 use gdk::Monitor;
 use glib::ObjectExt;
+use gtk::{gdk, glib};
 use itertools::Itertools;
 use once_cell::sync::Lazy;
 use simplexpr::{dynval::DynVal, SimplExpr};
 use std::{
     cell::RefCell,
     collections::{HashMap, HashSet},
+    marker::PhantomData,
     rc::Rc,
 };
 use tokio::sync::mpsc::UnboundedSender;
@@ -87,10 +88,6 @@ pub enum DaemonCommand {
 /// An opened window.
 #[derive(Debug)]
 pub struct EwwWindow {
-    /// Every window has an id, uniquely identifying it.
-    /// If no specific ID was specified whilst starting the window,
-    /// this will be the same as the window name.
-    pub instance_id: String,
     pub name: String,
     pub scope_index: ScopeIndex,
     pub gtk_window: Window,
@@ -111,11 +108,13 @@ impl EwwWindow {
     }
 }
 
-pub struct App<B> {
-    pub display_backend: B,
+pub struct App<B: DisplayBackend> {
     pub scope_graph: Rc<RefCell<ScopeGraph>>,
     pub eww_config: config::EwwConfig,
-    /// Map of all currently open windows by their IDs
+    /// Map of all currently open windows to their unique IDs
+    /// If no specific ID was specified whilst starting the window,
+    /// it will be the same as the window name.
+    /// Therefore, only one window of a given name can exist when not using IDs.
     pub open_windows: HashMap<String, EwwWindow>,
     pub instance_id_to_args: HashMap<String, WindowArguments>,
     /// Window names that are supposed to be open, but failed.
@@ -131,9 +130,10 @@ pub struct App<B> {
     pub window_close_timer_abort_senders: HashMap<String, futures::channel::oneshot::Sender<()>>,
 
     pub paths: EwwPaths,
+    pub phantom: PhantomData<B>,
 }
 
-impl<B> std::fmt::Debug for App<B> {
+impl<B: DisplayBackend> std::fmt::Debug for App<B> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.debug_struct("App")
             .field("scope_graph", &*self.scope_graph.borrow())
@@ -572,7 +572,6 @@ fn initialize_window<B: DisplayBackend>(
     window.show_all();
 
     Ok(EwwWindow {
-        instance_id: window_init.id.clone(),
         name: window_init.name.clone(),
         gtk_window: window,
         scope_index: window_scope,
@@ -626,6 +625,18 @@ fn get_gdk_monitor(identifier: Option<MonitorIdentifier>) -> Result<Monitor> {
     Ok(monitor)
 }
 
+/// Get the name of monitor plug for given monitor number
+/// workaround gdk not providing this information on wayland in regular calls
+/// gdk_screen_get_monitor_plug_name is deprecated but works fine for that case
+fn get_monitor_plug_name(display: &gdk::Display, monitor_num: i32) -> Option<&str> {
+    unsafe {
+        use glib::translate::ToGlibPtr;
+        let plug_name_pointer = gdk_sys::gdk_screen_get_monitor_plug_name(display.default_screen().to_glib_none().0, monitor_num);
+        use std::ffi::CStr;
+        CStr::from_ptr(plug_name_pointer).to_str().ok()
+    }
+}
+
 /// Returns the [Monitor][gdk::Monitor] structure corresponding to the identifer.
 /// Outside of x11, only [MonitorIdentifier::Numeric] is supported
 pub fn get_monitor_from_display(display: &gdk::Display, identifier: &MonitorIdentifier) -> Option<gdk::Monitor> {
@@ -643,7 +654,7 @@ pub fn get_monitor_from_display(display: &gdk::Display, identifier: &MonitorIden
         MonitorIdentifier::Name(name) => {
             for m in 0..display.n_monitors() {
                 if let Some(model) = display.monitor(m).and_then(|x| x.model()) {
-                    if model == *name {
+                    if model == *name || Some(name.as_str()) == get_monitor_plug_name(display, m) {
                         return display.monitor(m);
                     }
                 }