+++ crates/eww/src/display_backend.rs
@@ -1,5 +1,7 @@
 use crate::{widgets::window::Window, window_initiator::WindowInitiator};
 
+use gtk::gdk;
+
 #[cfg(feature = "wayland")]
 pub use platform_wayland::WaylandBackend;
 
@@ -8,6 +10,7 @@ pub use platform_x11::{set_xprops, X11Backend};
 
 pub trait DisplayBackend: Send + Sync + 'static {
     const IS_X11: bool;
+    const IS_WAYLAND: bool;
 
     fn initialize_window(window_init: &WindowInitiator, monitor: gdk::Rectangle, x: i32, y: i32) -> Option<Window>;
 }
@@ -16,6 +19,7 @@ pub struct NoBackend;
 
 impl DisplayBackend for NoBackend {
     const IS_X11: bool = false;
+    const IS_WAYLAND: bool = false;
 
     fn initialize_window(_window_init: &WindowInitiator, _monitor: gdk::Rectangle, x: i32, y: i32) -> Option<Window> {
         Some(Window::new(gtk::WindowType::Toplevel, x, y))
@@ -25,7 +29,9 @@ impl DisplayBackend for NoBackend {
 #[cfg(feature = "wayland")]
 mod platform_wayland {
     use crate::{widgets::window::Window, window_initiator::WindowInitiator};
+    use gtk::gdk;
     use gtk::prelude::*;
+    use gtk_layer_shell::LayerShell;
     use yuck::config::{window_definition::WindowStacking, window_geometry::AnchorAlignment};
 
     use super::DisplayBackend;
@@ -34,16 +40,17 @@ mod platform_wayland {
 
     impl DisplayBackend for WaylandBackend {
         const IS_X11: bool = false;
+        const IS_WAYLAND: bool = true;
 
         fn initialize_window(window_init: &WindowInitiator, monitor: gdk::Rectangle, x: i32, y: i32) -> Option<Window> {
             let window = Window::new(gtk::WindowType::Toplevel, x, y);
             // Initialising a layer shell surface
-            gtk_layer_shell::init_for_window(&window);
+            window.init_layer_shell();
             // Sets the monitor where the surface is shown
             if let Some(ident) = window_init.monitor.clone() {
                 let display = gdk::Display::default().expect("could not get default display");
                 if let Some(monitor) = crate::app::get_monitor_from_display(&display, &ident) {
-                    gtk_layer_shell::set_monitor(&window, &monitor);
+                    window.set_monitor(&monitor);
                 } else {
                     return None;
                 }
@@ -52,18 +59,18 @@ mod platform_wayland {
 
             // Sets the layer where the layer shell surface will spawn
             match window_init.stacking {
-                WindowStacking::Foreground => gtk_layer_shell::set_layer(&window, gtk_layer_shell::Layer::Top),
-                WindowStacking::Background => gtk_layer_shell::set_layer(&window, gtk_layer_shell::Layer::Background),
-                WindowStacking::Bottom => gtk_layer_shell::set_layer(&window, gtk_layer_shell::Layer::Bottom),
-                WindowStacking::Overlay => gtk_layer_shell::set_layer(&window, gtk_layer_shell::Layer::Overlay),
+                WindowStacking::Foreground => window.set_layer(gtk_layer_shell::Layer::Top),
+                WindowStacking::Background => window.set_layer(gtk_layer_shell::Layer::Background),
+                WindowStacking::Bottom => window.set_layer(gtk_layer_shell::Layer::Bottom),
+                WindowStacking::Overlay => window.set_layer(gtk_layer_shell::Layer::Overlay),
             }
 
             if let Some(namespace) = &window_init.backend_options.wayland.namespace {
-                gtk_layer_shell::set_namespace(&window, namespace);
+                window.set_namespace(namespace);
             }
 
             // Sets the keyboard interactivity
-            gtk_layer_shell::set_keyboard_interactivity(&window, window_init.backend_options.wayland.focusable);
+            window.set_keyboard_interactivity(window_init.backend_options.wayland.focusable);
 
             if let Some(geometry) = window_init.geometry {
                 // Positioning surface
@@ -83,27 +90,27 @@ mod platform_wayland {
                     AnchorAlignment::END => bottom = true,
                 }
 
-                gtk_layer_shell::set_anchor(&window, gtk_layer_shell::Edge::Left, left);
-                gtk_layer_shell::set_anchor(&window, gtk_layer_shell::Edge::Right, right);
-                gtk_layer_shell::set_anchor(&window, gtk_layer_shell::Edge::Top, top);
-                gtk_layer_shell::set_anchor(&window, gtk_layer_shell::Edge::Bottom, bottom);
+                window.set_anchor(gtk_layer_shell::Edge::Left, left);
+                window.set_anchor(gtk_layer_shell::Edge::Right, right);
+                window.set_anchor(gtk_layer_shell::Edge::Top, top);
+                window.set_anchor(gtk_layer_shell::Edge::Bottom, bottom);
 
                 let xoffset = geometry.offset.x.pixels_relative_to(monitor.width());
                 let yoffset = geometry.offset.y.pixels_relative_to(monitor.height());
 
                 if left {
-                    gtk_layer_shell::set_margin(&window, gtk_layer_shell::Edge::Left, xoffset);
+                    window.set_layer_shell_margin(gtk_layer_shell::Edge::Left, xoffset);
                 } else {
-                    gtk_layer_shell::set_margin(&window, gtk_layer_shell::Edge::Right, xoffset);
+                    window.set_layer_shell_margin(gtk_layer_shell::Edge::Right, xoffset);
                 }
                 if bottom {
-                    gtk_layer_shell::set_margin(&window, gtk_layer_shell::Edge::Bottom, yoffset);
+                    window.set_layer_shell_margin(gtk_layer_shell::Edge::Bottom, yoffset);
                 } else {
-                    gtk_layer_shell::set_margin(&window, gtk_layer_shell::Edge::Top, yoffset);
+                    window.set_layer_shell_margin(gtk_layer_shell::Edge::Top, yoffset);
                 }
             }
             if window_init.backend_options.wayland.exclusive {
-                gtk_layer_shell::auto_exclusive_zone_enable(&window);
+                window.auto_exclusive_zone_enable();
             }
             Some(window)
         }
@@ -115,6 +122,7 @@ mod platform_x11 {
     use crate::{widgets::window::Window, window_initiator::WindowInitiator};
     use anyhow::{Context, Result};
     use gdk::Monitor;
+    use gtk::gdk;
     use gtk::{self, prelude::*};
     use x11rb::protocol::xproto::ConnectionExt;
 
@@ -134,6 +142,7 @@ mod platform_x11 {
     pub struct X11Backend;
     impl DisplayBackend for X11Backend {
         const IS_X11: bool = true;
+        const IS_WAYLAND: bool = false;
 
         fn initialize_window(window_init: &WindowInitiator, _monitor: gdk::Rectangle, x: i32, y: i32) -> Option<Window> {
             let window_type =