+++ CHANGELOG.md
@@ -5,10 +5,17 @@ All notable changes to eww will be listed here, starting at changes since versio
 ## Unreleased
 
 ### Fixes
+- Re-enable some scss features (By: w-lfchen)
 - Fix and refactor nix flake (By: w-lfchen)
 - Fix remove items from systray (By: vnva)
+- Fix the gtk `stack` widget (By: ovalkonia)
+- Fix values in the `EWW_NET` variable (By: mario-kr)
+- Fix the gtk `expander` widget (By: ovalkonia)
+- Fix wayland monitor names support (By: dragonnn)
 
 ### Features
+- Update rust toolchain to 1.80.1 (By: w-lfchen)
+- Add `:fill-svg` and `:preserve-aspect-ratio` properties to images (By: hypernova7, w-lfchen)
 - Add `:truncate` property to labels, disabled by default (except in cases where truncation would be enabled in version `0.5.0` and before) (By: Rayzeq).
 - Add support for `:hover` css selectors for tray items (By: zeapoz)
 - Add scss support for the `:style` widget property (By: ovalkonia)